chore: 🤖 refact logger
