+++ packages/build-scripts/src/index.ts
@@ -1,8 +1,6 @@
 import Service from './Service';
 import Context, { createContext } from './Context';
 
-export type { ITaskConfig, IJestResult } from './Context';
-
 export * from './types';
 
 export {
