+++ packages/build-scripts/src/utils/loadPkg.ts
@@ -1,16 +1,16 @@
 import { resolve } from 'path';
 import { existsSync, readJsonSync } from 'fs-extra';
-import log from './log';
+import type { CreateLoggerReturns } from './logger';
 import type { Json } from '../types';
 
-const loadPkg = (rootDir: string): Json => {
+const loadPkg = (rootDir: string, logger: CreateLoggerReturns): Json => {
   const resolvedPath = resolve(rootDir, 'package.json');
   let config = {};
   if (existsSync(resolvedPath)) {
     try {
       config = readJsonSync(resolvedPath);
     } catch (err) {
-      log.info(
+      logger.info(
         'CONFIG',
         `Fail to load config file ${resolvedPath}, use empty object`,
       );
